requirements.txt
asgiref==3.2.10
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
